requirements.txt
appdirs==1.4.3
backports-abc==0.5
boto3==1.4.4
botocore==1.5.56
certifi==2017.4.17
chardet==3.0.3
click==6.7
cloudpickle==0.2.2
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.6
docutils==0.13.1
en-core-web-sm==1.2.0
Flask==0.12.2
ftfy==4.4.3
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gevent==1.2.1
greenlet==0.4.12
html5lib==0.999999999
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.0
murmurhash==0.26.4
numpy==1.12.1
packaging==16.8
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
pyparsing==2.2.0
python-crfsuite==0.9.2
python-dateutil==2.6.0
python-decouple==3.1
rasa-nlu==0.8.9
regex==2017.4.5
requests==2.16.0
s3transfer==0.1.10
scikit-learn==0.18.1
scipy==0.19.0
singledispatch==3.4.0.3
six==1.10.0
spacy==1.8.2
termcolor==1.1.0
thinc==6.5.2
toolz==0.8.2
tornado==4.5.1
tqdm==4.11.2
typing==3.6.1
ujson==1.35
urllib3==1.21.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wrapt==1.10.10
